chore(MWE-181): Dashing Bumblebe Nav icon fix
