+++ apps-ui/dashing-bumblebee-ui/app/contact/page.tsx
@@ -1,35 +1,28 @@
 'use client';
 
 import React from 'react';
-import { Field, Form, Formik } from 'formik';
+import { ContactCardContainer , CardData, contentRootAtomRW } from '@molitio/mwe-ui-core';
+import { useAtom } from 'jotai';
+
 
 export default function Page() {
-    const initialValues = {
-        firstName: '',
-    };
+    const [contentRoot] = useAtom(contentRootAtomRW);
+    const contentRootLeafs = contentRoot['home']?.leafs ?? {};
+    const cardTheTwoData: CardData[] = contentRootLeafs['contactCardCover']?.textContentCollection ?? [];
 
     return (
-        <div className="pt-40 h-screen">
-            <h1>Contact</h1>
-            <p>Send us a message!</p>
-            <div className="h-96">
-                <div>
-                    <Formik
-                        initialValues={initialValues}
-                        onSubmit={(values, actions) => {
-                            console.log({ values, actions });
-                            alert(JSON.stringify(values, null, 2));
-                            actions.setSubmitting(false);
-                        }}
-                    >
-                        <Form>
-                            <label htmlFor="firstName">First Name</label>
-                            <Field id="firstName" name="firstName" placeholder="First Name" />
-                            <button type="submit">Submit</button>
-                        </Form>
-                    </Formik>
+        <section className="hero min-h-screen bg-gray-800 items-start">
+            <div className="hero-content bg top-0 z-1 flex-col py-2">
+                <div className="container">
+                    <h2 className="text-white font-bold text-4xl align-items-center">
+                        Elérhetőségeink:
+                    </h2>
                 </div>
+                <div className="w-full mb-4 text-3xl py-2 rounded-box">
+                    <ContactCardContainer cards={cardTheTwoData} />
+                </div>
+               
             </div>
-        </div>
+        </section>
     );
 }
